module6/assignment/pyproject.toml
[project]
name = "module6-assignment"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.0",
    "pandas>=2.2.0",
    "websocket-client>=1.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
